requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
deepface==0.0.79
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.1
sqlite3
bcrypt==4.0.1
pandas==2.0.3
matplotlib==3.7.2
plotly==5.17.0
streamlit-authenticator==0.2.3
